Safari-611.1.10/LayoutTests/accessibility/aria-labelledby-stay-within-expected.txt

This tests that aria-labelledby does not append all sibling to an ARIA name

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS platformValueForW3CName(axButton) is "Reply Item Five"
PASS successfullyParsed is true

TEST COMPLETE
